third_party/blink/web_tests/wpt_internal/attribution-reporting/request-format.sub.https.html

<!doctype html>
<meta charset=utf-8>
<meta name=timeout content=long>
<meta name=variant content="?method=a&expected-eligible=navigation-source&expected-support=web">
<meta name=variant content="?method=img&expected-eligible=event-source, trigger&expected-support=web">
<meta name=variant content="?method=img&eligible&expected-eligible=event-source, trigger&expected-support=web">
<meta name=variant content="?method=open&expected-eligible=navigation-source&expected-support=web">
<meta name=variant content="?method=script&expected-eligible=event-source, trigger&expected-support=web">
<meta name=variant content="?method=script&eligible&expected-eligible=event-source, trigger&expected-support=web">
<meta name=variant content="?method=fetch">
<meta name=variant content="?method=fetch&eligible=event-source&expected-eligible=event-source&expected-support=web">
<meta name=variant content="?method=fetch&eligible=event-source&support=dummy&expected-eligible=event-source&expected-support=dummy">
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script src="resources/helpers.js"></script>
<script>
const searchParams = new URLSearchParams(location.search);
const expected_eligible =
    searchParams.get('expected-eligible') === null ? undefined : searchParams.get('expected-eligible');
const expected_support =
    searchParams.get('expected-support') === null ? undefined : searchParams.get('expected-support');

promise_test(async t => {
  const reg = await navigator.serviceWorker.register(
    'resources/attributionsrc-worker.js', {scope: './'});
  t.add_cleanup(() => reg.unregister());

  const sw = reg.installing || reg.waiting || reg.active;
  await wait_for_state(t, sw, 'activated');

  const p = new Promise(resolve => {
    navigator.serviceWorker.addEventListener('message',
      e => resolve(e.data), {once: true});
  });

  // Set mixed-case query params to ensure that they are propagated correctly.
  await registerAttributionSrc(t, {
    method: 'variant',
    extraQueryParams: {'aB': 'Cd'},
  });

  const message = await p;
  assert_equals(message.method, 'GET');
  assert_equals(message.headers['attribution-reporting-eligible'], expected_eligible);
  assert_equals(message.headers['attribution-reporting-support'], expected_support);
  assert_equals(message.referrer, location.toString());

  // TODO(apaseltiner): Test various referrer policies.
  // TODO(apaseltiner): Test cookie propagation.

  const expectedURL = blankURL();
  expectedURL.searchParams.set('aB', 'Cd');
  assert_equals(message.url, expectedURL.toString());
}, 'attributionsrc request has the proper format.');
</script>
